Move task processing to new file
